Addition of a Xataface web interface.
The content of the html directory has to be used with a recent
installation of Xataface to produce a web interface.
Tested with xataface-2.0.3 in /var/www/html/
Both conf.ini and index.php need to be adapted before use.
The default admin password, set in SAMADhi.sql is "ihDAMAS".
It can be changed from the interface itself.