Use das_client for importing CMS datasets
A valid certificate is now required to query DAS. In recent python
version (like the one used in CMSSW 8.0.x), it's not enough to pass the
key and cert options.

Instead of patching the existing code, we now use `das_client` for
performing the DAS query, which works as expected.